🚧 Separate DishSearch into container and presentational
